update leaderboard to use db ids
