+++ server/users.js
@@ -166,9 +166,10 @@ exports.getUsersList = function(callback) {
 }
 
 /* Return an array of users in the database, sorted by rank. */
-exports.getStudentsListSorted = function(lim, callback) {
+var getStudentsListSorted = function(lim, callback) {
     db.getStudentsListSorted(lim, callback);
 }
+exports.getStudentsListSorted = getStudentsListSorted;
 
 /*
  * Check if the account given by user and pass is valid.
@@ -318,5 +319,88 @@ exports.updateProfile = function (userId, request, callback) {
 
     db.updateUserByQuery(query, update, function (err,result) {
         return callback(err, result);
+    });
+}
+
+/*
+ * Fetch a list of students to display in the leaderboard.
+ * The list is sorted by decreasing points.
+ * Add rank to each student entry.
+ *
+ * If shrt is true, return leaderboard with max eight entries.
+ */
+exports.getLeaderboard = function(userid, shrt, callback) {
+    getStudentsListSorted(0, function(err, studentlist) {
+        if (err) {
+            logger.error(err);
+            return (err, []);
+        }
+
+        var rank = 0;
+        var last = -1;
+        var userind;
+
+        /* assign ranks to each student */
+        for (var i = 0; i < studentlist.length; ++i) {
+            /* subsequent users with same amount of points have same rank */
+            if (studentlist[i].points !== last) {
+                rank = i + 1;
+                last = studentlist[i].points;
+            }
+            studentlist[i].rank = rank;
+
+            if (studentlist[i]._id === userid)
+                userind = i;
+        }
+
+        if (shrt) {
+            if (studentlist.length < 8) {
+                callback(studentlist);
+                return;
+            }
+
+            var lb = [];
+            /* represents a '...' entry in the leaderboard table */
+            var dots = { rank: 0 };
+
+            /*
+             * If the user is in the top 6, display the top seven
+             * students, followed by "...".
+             */
+            if (userind < 6) {
+                for (var i = 0; i < 7; ++i)
+                    lb.push(studentlist[i]);
+                lb.push(dots);
+            } else {
+                /* The top 3 students are always displayed. */
+                lb.push(studentlist[0]);
+                lb.push(studentlist[1]);
+                lb.push(studentlist[2]);
+
+                /*
+                 * If the user is in the bottom four,
+                 * display the whole bottom four.
+                 */
+                if (userind >= studentlist.length - 4) {
+                    lb.push(dots);
+                    for (var i = studentlist.length - 4; i < studentlist.length; ++i)
+                        lb.push(studentlist[i]);
+                } else {
+                    /*
+                     * Otherwise, display the user with the
+                     * students directly above and below them.
+                     */
+                    lb.push(dots);
+                    lb.push(studentlist[userind - 1]);
+                    lb.push(studentlist[userind]);
+                    lb.push(studentlist[userind + 1]);
+                    lb.push(dots);
+                }
+            }
+
+            callback(lb);
+        } else {
+            callback(studentlist);
+        }
     });
 }
